requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.6.0
Django==3.1.7
django-advanced-filters==1.3.0
django-allauth==0.44.0
django-autoslug==1.9.8
django-braces==1.14.0
django-cors-headers==3.7.0
django-filter==1.1.0
django-money==1.3.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-tinymce4-lite==1.8.0
djangorestframework==3.12.2
djangorestframework-filters==0.11.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
jsmin==2.2.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
py-moneyed==0.8.0
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
